requirements.txt
asgiref==3.4.1
boto3==1.17.61
botocore==1.20.112
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==3.3
jmespath==0.10.0
mysqlclient==2.0.3
Pillow==8.4.0
psycopg2==2.9.2
python-dateutil==2.8.2
pytz==2021.3
razorpay==1.2.0
requests==2.26.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
